+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 
 ### Fixed
 - JavaScript source maps on Chrome/Safari [#407]
+- Nunjucks cache for the imported templates.
 
 ## [1.16.1] - 2023-03-29
 ### Added